requirements.txt
pyspark==3.5.0 # Para desenvolvimento e teste local
requests==2.31.0 # Para download programático dos dados
delta-spark==3.5.0 # Suporte ao Delta Lake no PySpark